+++ CHANGELOG.md
@@ -6,8 +6,10 @@
 
 - Add inspection and quick fix for unused local variables.
 - Add varchook to valid types.
+- Implement support for enum_getreverseindex command.
 
 ### Changed
+
 - Update to IntelliJ 2024.1.
 
 ## [1.5.0] - 2023-11-27